Add data set which can construct our testing data.
We assume the CPU byte address is that :

|  tag  |  index  |  byte offset |

tag : Tag can detect the  data whather or not is turely in cache.
index : Index can show us that CPU want to store which cache entry.
byre offset : Which byte block need to store in the data of cache.

When CPU want to store cache, we can use this method to know the information