Add Direct mapped Cache and CPU address.
When CPU want to store data in memory. We can observe CPU address to know some information.

|  tag  |  index  |  byte offset |

tag : Tag can detect the  data whather or not is turely in cache.
index : Index can show us that CPU want to store which cache entry.
byre offset : Which byte block need to store in the data of cache.

---------------------------

Cache information :

index : 0 -> don't have data in this entry; 1 -> have data in this entry.
tag : Tag can detect the  data whather or not is turely in cache.
data : PhoneBook LastName.

----------------------------

When memory want to store the data in the cache, the mapping function :
(memory block address ) mod ( number of cache blocks in the cache)

---------------------------

the test_address.txt can work, but address.txt can not work.